move JS errors-related code from Selenide and Navigator to a dedicated class JavascriptErrorsCollector
